+++ src/create_wce_objects.h
@@ -77,7 +77,7 @@ namespace wincalc
     std::shared_ptr<SingleLayerOptics::SpecularLayer>
       create_specular_layer(std::shared_ptr<wincalc::Product_Data_Optical> const & product_data,
                             window_standards::Optical_Standard_Method const & method,
-                            int number_of_layers,
+                            size_t number_of_layers,
                             Spectal_Data_Wavelength_Range_Method const & type =
                               Spectal_Data_Wavelength_Range_Method::FULL,
                             int number_visible_bands = 5,
@@ -96,7 +96,7 @@ namespace wincalc
     std::shared_ptr<SingleLayerOptics::CBSDFLayer>
       create_bsdf_layer(std::shared_ptr<wincalc::Product_Data_Optical> const & product_data,
                         window_standards::Optical_Standard_Method const & method,
-                        int number_of_layers,
+                        size_t number_of_layers,
                         SingleLayerOptics::CBSDFHemisphere const & bsdf_hemisphere,
                         Spectal_Data_Wavelength_Range_Method const & type =
                           Spectal_Data_Wavelength_Range_Method::FULL,
